refactor: Photo 상위 엔티티 domain.model 패키지로 이동
