+++ backend/forgather/src/test/java/com/forgather/fake/FakeContentStorage.java
@@ -1,14 +1,11 @@
 package com.forgather.fake;
 
-import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.net.URL;
 import java.util.List;
 
 import org.springframework.web.multipart.MultipartFile;
 
-import com.forgather.domain.space.model.Photo;
+import com.forgather.domain.model.Photo;
 import com.forgather.domain.upload.domain.ContentsStorage;
 
 /**
@@ -18,21 +15,6 @@ public class FakeContentStorage implements ContentsStorage {
     @Override
     public String upload(String spaceCode, MultipartFile file) throws IOException {
         return "";
-    }
-
-    @Override
-    public InputStream download(String photoPath) {
-        return null;
-    }
-
-    @Override
-    public File downloadSelected(String tempPath, String spaceCode, List<String> photoPaths) {
-        return null;
-    }
-
-    @Override
-    public URL issueDownloadUrl(String photoPath) {
-        return null;
     }
 
     @Override
